AccountMicroService/requirement.txt
psycopg2-binary
peewee
grpcio
grpcio-tools